# backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Blockchain Gateway Backend API
  version: 3.0.0
  description: |
    Backend API for chaincode lifecycle orchestration with RBAC.
    
    This API provides comprehensive management of:
    - User authentication and authorization with JWT tokens
    - Role-based access control (ADMIN, ORG_ADMIN, USER, VIEWER)
    - Chaincode lifecycle management (upload, validate, approve, deploy)
    - Deployment orchestration with workflow management
    - Certificate synchronization with Fabric CA
    - Audit logging for all operations
    
    ## Authentication
    
    The API uses JWT tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Roles and Permissions
    
    - **ADMIN**: Full system access, can manage users, approve chaincodes, configure system
    - **ORG_ADMIN**: Organization administrator, can upload/deploy chaincodes, view users
    - **USER**: Regular user, can invoke/query chaincodes, manage assets
    - **VIEWER**: Read-only access, can query chaincodes and view assets
    
  contact:
    name: Blockchain Gateway Team
    email: support@blockchain.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.blockchain.com/api/v1
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with username and password
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type:
                    type: string
                    example: bearer
                  expires_in:
                    type: integer
                    example: 1800
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user info
      description: Get information about the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Management endpoints
...
  /certificates/sync:
    post:
      tags:
        - Certificate Management
      summary: Synchronize certificates
      description: Synchronize user certificates with Fabric CA (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                force_sync:
                  type: boolean
                  default: false
                  description: Force immediate sync instead of background
      responses:
        '200':
          description: Certificate synchronization completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateSyncResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /certificates/register:
    post:
      tags:
        - Certificate Management
      summary: Register user with Fabric CA
      description: Register a new user with Fabric CA (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateRegisterResponse'
        '400':
          description: Invalid registration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /certificates/enroll:
    post:
      tags:
        - Certificate Management
      summary: Enroll user with Fabric CA
      description: Enroll user with Fabric CA to get certificate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateEnrollRequest'
      responses:
        '200':
          description: User enrolled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateEnrollResponse'
        '400':
          description: Invalid enrollment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /certificates/revoke:
    post:
      tags:
        - Certificate Management
      summary: Revoke certificate
      description: Revoke a certificate in Fabric CA (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CertificateRevokeRequest'
      responses:
        '200':
          description: Certificate revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Invalid revocation data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /certificates/status/{certificate_id}:
    get:
      tags:
        - Certificate Management
      summary: Get certificate status
      description: Get certificate status (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: certificate_id
          in: path
          required: true
          description: Certificate ID
          schema:
            type: string
      responses:
        '200':
          description: Certificate status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateStatusResponse'
        '404':
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /certificates/my-certificate:
    get:
      tags:
        - Certificate Management
      summary: Get my certificate status
      description: Get current user's certificate status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Certificate status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificateStatusResponse'
        '404':
          description: No certificate found for current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        detail:
          type: string
          example: "Detailed error information"

    # Authentication schemas
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 1800
    UserInfoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: "admin"
        email:
          type: string
          example: "admin@blockchain.com"
        role:
          type: string
          enum: [ADMIN, ORG_ADMIN, USER, VIEWER]
          example: "ADMIN"
        organization:
          type: string
          example: "Blockchain Gateway"
        msp_id:
          type: string
          example: "Org1MSP"
        is_active:
          type: boolean
          example: true
        last_login:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"

    # User Management schemas
...
    CertificateRegisterRequest:
      type: object
      required:
        - username
        - organization
      properties:
        username:
          type: string
          example: "john_doe"
        organization:
          type: string
          example: "org1"
        role:
          type: string
          default: "user"
          example: "user"
    CertificateRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User registered with Fabric CA"
        data:
          type: object
          properties:
            username:
              type: string
              example: "john_doe"
            secret:
              type: string
              example: "secret123"
            certificate_id:
              type: string
              example: "cert123456789"
    CertificateEnrollRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "john_doe"
        password:
          type: string
          example: "secret123"
        organization:
          type: string
          default: "org1"
          example: "org1"
    CertificateEnrollResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User enrolled successfully"
        data:
          type: object
          properties:
            certificate_id:
              type: string
              example: "cert123456789"
            certificate:
              type: string
              example: "-----BEGIN CERTIFICATE-----\n..."
            private_key:
              type: string
              example: "-----BEGIN PRIVATE KEY-----\n..."
    CertificateRevokeRequest:
      type: object
      required:
        - certificate_id
      properties:
        certificate_id:
          type: string
          example: "cert123456789"
        reason:
          type: string
          default: "unspecified"
          example: "User left organization"
    CertificateStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            certificate_id:
              type: string
              example: "cert123456789"
            user_id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            username:
              type: string
              example: "john_doe"
            organization:
              type: string
              example: "org1"
            msp_id:
              type: string
              example: "Org1MSP"
            is_active:
              type: boolean
              example: true
            status:
              type: string
              example: "active"
            last_sync:
              type: string
              format: date-time
              nullable: true
              example: "2023-12-01T10:30:00Z"
    CertificateSyncResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Certificate synchronization completed"
        results:
          type: object
          properties:
            total_users:
              type: integer
              example: 25
            valid_certificates:
              type: integer
              example: 23
            invalid_certificates:
              type: integer
              example: 2
            sync_errors:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  error:
                    type: string
              example: []